+++ tests/integration/test_credit_card.py
@@ -162,14 +162,16 @@ class TestCreditCard(unittest.TestCase):
             "customer_id": customer.id,
             "number": "4000111111111115",
             "expiration_date": "05/2014",
-            "options": {"verify_card": True}
+            "options": {"verify_card": True},
+            "device_session_id": "abc123"
         })
 
         self.assertFalse(result.is_success)
         verification = result.credit_card_verification
         self.assertIsInstance(verification.risk_data, RiskData)
-        self.assertEqual(None, verification.risk_data.id)
+        self.assertTrue(hasattr(verification.risk_data, 'id'))
         self.assertEqual("Not Evaluated", verification.risk_data.decision)
+        self.assertTrue(hasattr(verification.risk_data, 'device_data_captured'))
 
     def test_successful_create_with_card_verification_returns_risk_data(self):
         customer = Customer.create().customer
@@ -177,14 +179,16 @@ class TestCreditCard(unittest.TestCase):
             "customer_id": customer.id,
             "number": "4111111111111111",
             "expiration_date": "05/2014",
-            "options": {"verify_card": True}
+            "options": {"verify_card": True},
+            "device_session_id": "abc123"
         })
 
         self.assertTrue(result.is_success)
         verification = result.credit_card.verification
         self.assertIsInstance(verification.risk_data, RiskData)
-        self.assertEqual(None, verification.risk_data.id)
+        self.assertTrue(hasattr(verification.risk_data, 'id'))
         self.assertEqual("Not Evaluated", verification.risk_data.decision)
+        self.assertTrue(hasattr(verification.risk_data, 'device_data_captured'))
 
     def test_create_with_card_verification(self):
         customer = Customer.create().customer