+++ braintree/payment_method.py
@@ -61,6 +61,8 @@ class PaymentMethod(Resource):
             "expiration_year",
             "number",
             "payment_method_nonce",
+            "paypal_refresh_token",
+            "paypal_vault_without_upgrade",
             "token",
             {
                 "billing_address": Address.create_signature()