+++ braintree/webhook_notification.py
@@ -12,6 +12,7 @@ from braintree.validation_error_collection import ValidationErrorCollection
 from braintree.connected_merchant_paypal_status_changed import ConnectedMerchantPayPalStatusChanged
 from braintree.connected_merchant_status_transitioned import ConnectedMerchantStatusTransitioned
 from braintree.ideal_payment import IdealPayment
+from braintree.granted_payment_instrument_update import GrantedPaymentInstrumentUpdate
 
 class WebhookNotification(Resource):
     class Kind(object):
@@ -41,6 +42,7 @@ class WebhookNotification(Resource):
         DisputeWon = "dispute_won"
         IdealPaymentComplete = "ideal_payment_complete"
         IdealPaymentFailed = "ideal_payment_failed"
+        GrantedPaymentInstrumentUpdate = "granted_payment_instrument_update"
 
     @staticmethod
     def parse(signature, payload):
@@ -78,6 +80,8 @@ class WebhookNotification(Resource):
             self.account_updater_daily_report = AccountUpdaterDailyReport(gateway, node_wrapper['account_updater_daily_report'])
         elif "ideal_payment" in node_wrapper:
             self.ideal_payment = IdealPayment(gateway, node_wrapper['ideal_payment'])
+        elif "granted_payment_instrument_update" in node_wrapper:
+            self.granted_payment_instrument_update = GrantedPaymentInstrumentUpdate(gateway, node_wrapper["granted_payment_instrument_update"])
 
         if "errors" in node_wrapper:
             self.errors = ValidationErrorCollection(node_wrapper['errors'])