+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add braintree.Dispute.remove_evidence
 * Add braintree.Dispute.search
 * Add braintree.DocumentUpload
+* Add `device_data_captured` field to `risk_data`
 
 ## 3.37.2
 * Fix a bug where a null value for `amount` in `CreditCardVerification` would result in a `ValueError`