+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 * Add support for upgrading a PayPal future payment refresh token to a billing agreement
 * Add logic to remove deprecation warnings for encodestring and decodestring when used with python 3 (#92)
+* Fix spec to expect PayPal transaction to settle immediately after successful capture
+* Add GrantedPaymentInstrumentUpdate webhook support
 
 ## 3.38.0
 * Add iDEAL webhook support