+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 3.39.0
 * Add support for upgrading a PayPal future payment refresh token to a billing agreement
 * Fix braintree.Dispute.search to take a list of search criteria
 * Add logic to remove deprecation warnings for encodestring and decodestring when used with python 3 (#92)
@@ -7,6 +7,8 @@
 * Add ability to create a transaction from a shared nonce
 * Add `options` -> `paypal` -> `shipping` for creating & updating customers as well as creating payment methods
 * Do not convert to Decimal if amount is None in AuthorizationAdjustement (#70)
+* Add `device_data_captured` field to `risk_data`
+* Add `bin_data` to `payment_method_nonce`
 
 ## 3.38.0
 * Add iDEAL webhook support
@@ -24,7 +26,6 @@
 * Add braintree.Dispute.remove_evidence
 * Add braintree.Dispute.search
 * Add braintree.DocumentUpload
-* Add `device_data_captured` field to `risk_data`
 
 ## 3.37.2
 * Fix a bug where a null value for `amount` in `CreditCardVerification` would result in a `ValueError`